updates to structure of the toolkit, initial functionality added to the CLI arg parser
