--- splicing_toolkit/main.py
@@ -1,3 +1,0 @@
-from utils import splicing_classifier
-
-splicing_classifier('LRS_url', 'output_dir')
